Changed template methods in ModelFactory class.

+++ Basic/ModelFactory.cpp
@@ -6,10 +6,10 @@ ModelFactory::ModelFactory()
 
 ModelFactory::~ModelFactory()
 {
-	for (int i = 0; i < vaos.size(); i++)
+	for (size_t i = 0; i < vaos.size(); i++)
 		glDeleteVertexArrays(1, &vaos[i]);
 
-	for (int i = 0; i < vbos.size(); i++)
+	for (size_t i = 0; i < vbos.size(); i++)
 		glDeleteBuffers(1, &vbos[i]);
 }
